update all actions
