+++ examples/demo/src/button.pc
@@ -8,17 +8,20 @@
 
 
 
+
+
+
+
 <!--
-  @frame { width: 1024, height: 768, x: 386, y: 24 }
+  @frame { width: 1024, height: 768, x: -1508, y: 102 }
 -->
-<div>
-  <mod.TEST />
-</div>
-
-
+<span>
+  Something like that
+</span>
 <!--
-  @frame { width: 1024, height: 768, x: 58, y: 176 }
+  @frame { width: 1024, height: 768, x: -114, y: 249 }
 -->
 <span>
   Something like this
 </span>
+
