+++ packages/paperclip-visual-editor/src/components/Editor/Center/Canvas/Tools/Selectable/index.tsx
@@ -11,7 +11,6 @@ import * as styles from "./index.pc";
 import { Dispatch } from "redux";
 import {
   Action,
-  canvasMouseUp,
   resizerMoved,
   resizerPathMoved,
   resizerPathStoppedMoving,
